requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
et-xmlfile==1.1.0
fastapi==0.75.0
h11==0.13.0
httptools==0.4.0
idna==3.3
numpy==1.22.3
openai==0.16.0
openpyxl==3.0.9
pandas==1.4.1
pandas-stubs==1.2.0.53
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8
websockets==10.2
